03 : Using JSX
another messy recap : Today I learned about JSX, I learned it's syntactic
sugar for the React raw API, it's puporse is to make reading and writing
easier and more intuitive. It is very reminiscent of HTML, though it has
differences. Since JSX is not js, we need Babel to convert it for us.
Interpolation is similar to js, but we use only {} and nothing else. We
can use props by using the spread operator, for now inside a self
inclosing div. We can also combine other attributes with our props, the
order we declare them is important, the last one is the one that will get
'used(?)'. Also read a bit on the difference between statement and
expression in js, an expression produces a value, roughly, a statement
performs an action. Though I still need to read more on that.