04: Creating custom components Components in React are basically functions which return something that is "renderable".
Using a function with create element
We can pass a function to createElement, that means, we can create a
function that returns an element, and use the createElement to create
new elements that inherit from that function and pass their specifics
(like children, class etc). Like a template?

Creating a component with JSX
But we can achieve the same by creating a component and using JSX. In
order to tell Babel that our component name is our function name, we need
to capitalize it (if not it will be read as a string).

Tracking errors with prop-types
In development, we can use the prop-types package validators to make sure
the data we receive is valid. E.g: Here an error will be raised because
we tried to pass a number to our greeting, and PropTypes is checking for
us if it is a string. Another one will be raised because we didn't pass
a subject, and we told PropTypes a value is required.

Using Fragments
Lastly, if we want to pass something directly to the root, not as a div
inside of the root, if we need side-by-side elements, instead of nested,
we use React Fragments.