06: Forms The first exercise was simple: I conected the label and input by giving the label a htmlFor with the same name as the input id, we do that so focus will work, and it will be accessible to e-readers. Also, by giving the input an id we can access the input value more directly.
In the first extra credit we used useRef() to achieve the same end. We
create a ref, assign it and access it. useRef is like a "box"that can
hold a mutable value in it's 'current'property.

The second extra credit I had a tougher time. This is the first version.
It somewhat works. I tried usind getElementById to get the input, but it
hadn't been rendered to the DOM, so I kept getting null. I used the
useEffect hook because it runs after the component has been mounted. This
way I was able to get the input by the ID, listen to the keyUp, and check
if what was typed was in lowercase, if in uppercase, I set an alert.
HOWEVER: Kent asks us to use the useState hook and setError. I think it
will actually look cleaner.