06: Forms We create a hanleChange function that we listen on the input. When I first tried, I actually selected the input by the id and added an event listener. Because it hadn't been rendered to the DOM, I put it inside an useEffect hook, since that runs after the component has been mounted. UGH. Way more complicated than it need be.
We can actually listen on the input tag and pass it the handleChange.
We pass the handleChange the event, and from that we can get the value.
Kent uses destructuring to get the target.value directly. Now that we
have the value we must check if it's lowercase.

We need to show the user an error when they enter an upper case
letter, to do that we need to maintain some error state, and to
do that, we use another React Hook: React.useState

The initial value for our state is null, it returns this.state.error
and this.setError. So new state gives us an array, where the first
value is the state value over time, and the second a mechanism for
updating that state. With that, we can setError, check if it's valid,
and return whatever accordingly.

Then, we can display an error, and disable our submit button.
(obs.: the div is rendered whether there is an error or not, but
since there is no error, nothing is displayed. we would have to
add some control flow for it to not be rendered at all).
We do the Boolean(error) to make sure we get a true or false value.