07: rendering arrays / the key prop Without a key prop with an unique consistent id, React can't keep track of items being rendered from an array. E.g., if you have a form with a label and input, and an remove buttom for each input, if you change the value in the input, and then remove the whole thing, React gets the values of the elements inputs mixed up.
With a key prop with an unique consistent id, React can keep track
of where everything should go, and we don't get those pesky bugs.

This problem will happen not only to inputs, but to any component that
needs to maintain state over time. We use the key prop so React can
associate the changes to the corect item.